project4

+++ heroList.cpp
@@ -239,7 +239,7 @@ void HeroList::addLoser(HeroNode *&loserTeam) {
 ** Description:     calculates health to restore
 *********************************************************************/
 int HeroList::restoreHealth(const HeroList::HeroNode *player) {
-    double restoreHealth = 0;
+    int restoreHealth = 0;
 
     if ( player->hero->getType() == "Vampire" ) {
         // max strength is 18
@@ -249,7 +249,7 @@ int HeroList::restoreHealth(const HeroList::HeroNode *player) {
         cout << "Vampire Health Lost " << strengthLost << endl;
 
         restoreHealth = ((rand() % 10 + 1) / 10.0) * strengthLost;
-        cout << "Random num generated " << ((rand() % 10 + 1) / 10) << endl;
+        cout << "Random num generated " << ((rand() % 10 + 1) / 10.0) << endl;
         cout << "Health to restore " << restoreHealth << endl;
 
         restoreHealth += player->hero->getStrength();
@@ -263,7 +263,7 @@ int HeroList::restoreHealth(const HeroList::HeroNode *player) {
         cout << "Barbarian Health Lost " << strengthLost << endl;
 
         restoreHealth = ((rand() % 10 + 1) / 10.0) * strengthLost;
-        cout << "Random num generated " << ((rand() % 10 + 1) / 10) << endl;
+        cout << "Random num generated " << ((rand() % 10 + 1) / 10.0) << endl;
         cout << "Health to restore " << restoreHealth << endl;
 
         restoreHealth += player->hero->getStrength();
@@ -277,7 +277,7 @@ int HeroList::restoreHealth(const HeroList::HeroNode *player) {
         cout << "Blue Men Health Lost " << strengthLost << endl;
 
         restoreHealth = ((rand() % 10 + 1) / 10.0) * strengthLost;
-        cout << "Random num generated " << ((rand() % 10 + 1) / 10) << endl;
+        cout << "Random num generated " << ((rand() % 10 + 1) / 10.0) << endl;
         cout << "Health to restore " << restoreHealth << endl;
 
         restoreHealth += player->hero->getStrength();
@@ -291,7 +291,7 @@ int HeroList::restoreHealth(const HeroList::HeroNode *player) {
         cout << "Medusa Health Lost " << strengthLost << endl;
 
         restoreHealth = ((rand() % 10 + 1) / 10.0) * strengthLost;
-        cout << "Random num generated " << ((rand() % 10 + 1) / 10) << endl;
+        cout << "Random num generated " << ((rand() % 10 + 1) / 10.0) << endl;
         cout << "Health to restore " << restoreHealth << endl;
 
         restoreHealth += player->hero->getStrength();
@@ -305,7 +305,7 @@ int HeroList::restoreHealth(const HeroList::HeroNode *player) {
         cout << "Harry Potter Health Lost " << strengthLost << endl;
 
         restoreHealth = ((rand() % 10 + 1) / 10.0) * strengthLost;
-        cout << "Random num generated " << ((rand() % 10 + 1) / 10) << endl;
+        cout << "Random num generated " << ((rand() % 10 + 1) / 10.0) << endl;
         cout << "Health to restore " << restoreHealth << endl;
 
         restoreHealth += player->hero->getStrength();
